refactor: remove deprecations around score and constraints (#2106)

+++ core/src/test/java/ai/timefold/solver/core/impl/score/stream/common/inliner/HardSoftScoreInlinerTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.Collections;
 import java.util.Map;
 
-import ai.timefold.solver.core.api.score.buildin.hardsoft.HardSoftScore;
+import ai.timefold.solver.core.api.score.HardSoftScore;
 import ai.timefold.solver.core.api.score.stream.Constraint;
 import ai.timefold.solver.core.impl.domain.solution.descriptor.SolutionDescriptor;
 import ai.timefold.solver.core.impl.score.constraint.ConstraintMatchPolicy;
@@ -95,13 +95,13 @@ class HardSoftScoreInlinerTest extends AbstractScoreInlinerTest<TestdataHardSoft
     void impactAllMatchWeightOverflow() {
         var constraintWeight = HardSoftScore.of(10, 100);
         var impacter = buildScoreImpacter(constraintWeight);
-        assertThatThrownBy(() -> impacter.impactScore(Integer.MAX_VALUE, ConstraintMatchSupplier.empty()))
+        assertThatThrownBy(() -> impacter.impactScore(Long.MAX_VALUE, ConstraintMatchSupplier.empty()))
                 .isInstanceOf(ArithmeticException.class);
     }
 
     @Test
     void impactAllTotalOverflow() {
-        var constraintWeight = HardSoftScore.of(Integer.MAX_VALUE, Integer.MAX_VALUE);
+        var constraintWeight = HardSoftScore.of(Long.MAX_VALUE, Long.MAX_VALUE);
         var impacter = buildScoreImpacter(constraintWeight);
         impacter.impactScore(1, ConstraintMatchSupplier.empty()); // This will send the total right to the limit.
         assertThatThrownBy(() -> impacter.impactScore(1, ConstraintMatchSupplier.empty()))
